pyproject.toml
[project]
name = "spralingua-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.2.1",
]
